Moved username and key to capabilities + tests
