+++ testing/test_saucelabs.py
@@ -5,6 +5,7 @@
 from functools import partial
 import os
 import datetime
+import json
 
 import pytest
 
@@ -75,6 +76,83 @@ def test_invalid_credentials_file(failure, monkeypatch, tmpdir):
     out = failure()
     messages = ['Sauce Labs Authentication Error', 'basic auth failed']
     assert any(message in out for message in messages)
+
+
+def test_no_sauce_options(monkeypatch, testdir):
+    username = 'foo'
+    access_key = 'bar'
+    monkeypatch.setenv('SAUCELABS_USERNAME', username)
+    monkeypatch.setenv('SAUCELABS_API_KEY', access_key)
+
+    capabilities = {'browserName': 'chrome'}
+    variables = testdir.makefile('.json', '{{"capabilities": {}}}'.format(
+        json.dumps(capabilities)))
+
+    file_test = testdir.makepyfile("""
+        import pytest
+        @pytest.mark.nondestructive
+        def test_sauce_capabilities(driver_kwargs):
+            assert driver_kwargs['desired_capabilities']['username'] == '{}'
+            assert driver_kwargs['desired_capabilities']['accessKey'] == '{}'
+            try:
+                driver_kwargs['desired_capabilities']['sauce:options']
+                raise AssertionError('<sauce:options> should not be present!')
+            except KeyError:
+                pass
+    """.format(username, access_key))
+
+    testdir.quick_qa(
+        '--driver', 'saucelabs', '--variables',
+        variables, file_test, passed=1)
+
+
+def test_empty_sauce_options(monkeypatch, testdir):
+    capabilities = {'browserName': 'chrome'}
+    expected = {'name': 'test_empty_sauce_options.test_sauce_capabilities'}
+    run_sauce_test(capabilities, expected, monkeypatch, testdir)
+
+
+def test_merge_sauce_options(monkeypatch, testdir):
+    version = {'seleniumVersion': '3.8.1'}
+    capabilities = {'browserName': 'chrome', 'sauce:options': version}
+    expected = {'name': 'test_merge_sauce_options.test_sauce_capabilities'}
+    expected.update(version)
+    run_sauce_test(capabilities, expected, monkeypatch, testdir)
+
+
+def test_merge_sauce_options_with_conflict(monkeypatch, testdir):
+    name = 'conflict'
+    capabilities = {'browserName': 'chrome', 'sauce:options': {'name': name}}
+    expected = {'name': name}
+    run_sauce_test(capabilities, expected, monkeypatch, testdir)
+
+
+def run_sauce_test(capabilities, expected_result, monkeypatch, testdir):
+    username = 'foo'
+    access_key = 'bar'
+
+    monkeypatch.setenv('SAUCELABS_USERNAME', username)
+    monkeypatch.setenv('SAUCELABS_API_KEY', access_key)
+    monkeypatch.setenv('SAUCELABS_W3C', 'true')
+
+    expected_result.update({'username': username,
+                            'accessKey': access_key,
+                            'tags': ['nondestructive']})
+
+    variables = testdir.makefile('.json', '{{"capabilities": {}}}'.format(
+        json.dumps(capabilities)))
+
+    file_test = testdir.makepyfile("""
+        import pytest
+        @pytest.mark.nondestructive
+        def test_sauce_capabilities(driver_kwargs):
+            actual = driver_kwargs['desired_capabilities']['sauce:options']
+            assert actual == {}
+    """.format(expected_result))
+
+    testdir.quick_qa(
+        '--driver', 'saucelabs', '--variables',
+        variables, file_test, passed=1)
 
 
 @pytest.mark.parametrize(('auth_type', 'auth_token'),
